+++ camel/agents/chat_agent.py
@@ -25,7 +25,7 @@ from camel.agents import BaseAgent
 from camel.configs import BaseConfig, ChatGPTConfig
 from camel.functions import OpenAIFunction
 from camel.memories import BaseMemory, ChatHistoryMemory, MemoryRecord
-from camel.memories.context_creator.default import DefaultContextCreator
+from camel.memories.context_creators.default import DefaultContextCreator
 from camel.messages import BaseMessage, FunctionCallingMessage, OpenAIMessage
 from camel.models import BaseModelBackend, ModelFactory
 from camel.responses import ChatAgentResponse
