+++ camel/memories/context_creators/base.py
