+++ camel/storages/dict_storages/base.py
